% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{Constr_net}
\alias{Constr_net}
\title{Constructing cell-type-specific GRN for single-cell transcriptomics data}
\usage{
Constr_net(
  object,
  select_ratio = 0.5,
  n_selection = 10,
  cut_ratio = 0.95,
  keep_ratio = 0.25,
  min_cells = 25,
  n.core = NULL,
  n.core_cp = 4,
  use_py = F,
  env = "base"
)
}
\arguments{
\item{object}{scRank object generated from \code{\link{CreateScRank}}.}

\item{select_ratio}{percent cells of total cells to be selected in random selection. Default is 0.5.}

\item{n_selection}{number of selection to create \code{n_selection} networks for a cell type. Default is 10.}

\item{cut_ratio}{threshold of top percent edge weight to be cut. Edge with weight lower than the threshold will be cut. Default is 0.95.}

\item{keep_ratio}{percent of the weaker edge weight should be kept. Default is 0.25.}

\item{min_cells}{the minimum number of cells in a particular cell type retained for constructing robust scGRN. Cell type with cells lower than \code{min_cells} will be droped. Default is 25.}

\item{n.core}{Number of CPU cores to use for constructing network. Default is all cores - 1.}

\item{n.core_cp}{Number of CPU cores to use for integrating network. Default is 4. Recommendation is the number lower than the number of cell type.}

\item{use_py}{Whether to use python to conduct CP decomposition}

\item{env}{When using python in \code{use_py}, please define the python environment of installed tensorly and hdf5. Default is the 'base' environment. Anaconda is recommended.}
}
\value{
scRank object containing the gene regulatory network for every cell type.
}
\description{
construct gene-gene co-expression adjacency matrix for every cell type with regression strategy.
}
