% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{CreateScRank}
\alias{CreateScRank}
\title{scRank object}
\usage{
CreateScRank(
  input,
  meta = NULL,
  cell_type = "celltype",
  species,
  drug = NULL,
  target = NULL,
  type = "antagonist",
  if_cluster = F,
  var.gene = NULL
)
}
\arguments{
\item{input}{gene expression profile formatted by matrix, data frame, or \code{Seurat}}

\item{meta}{meta data describing the cells in input and its corresponding cell type; can be left as \code{NULL} if \code{input} is a \code{Seurat} object}

\item{cell_type}{the name of column which containing cell type labels for
each cell corresponding to gene-by-cell expression matrix}

\item{species}{characters meaning species of the single-cell transcriptomic data. \code{human} or \code{mouse}.}

\item{drug}{characters meaning the name of drug.}

\item{target}{characters meaning the gene encoded direct target of interested drug.}

\item{type}{characters meaning the MOAs of drug including antagonist or agonist. Default is antagonist.}

\item{if_cluster}{A logical meaning whether clustering single-cell transcriptomic data. Default is \code{FALSE}.}

\item{var.genes}{var.genes (optional) vector of gene names representing the subset of genes used for clustering.}
}
\value{
scRank object
}
\description{
create scRank object using single-cell transcriptomic data.
}
