% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_drug_function}
\alias{plot_drug_function}
\title{Plotting drug function on cell type}
\usage{
plot_drug_function(
  object,
  celltype = NULL,
  category = "H",
  top_number = 5,
  show_leading_edge = T
)
}
\arguments{
\item{object}{scRank object generated from \code{\link{Constr_net}}}

\item{celltype}{select a cell type to measure the drug curative efficacy.}

\item{category}{the category of gene set in MSigDB. Default is "H" representing the hallmark gene sets.}

\item{top_number}{the number of top pathways to be shown in the plot. Default is 5.}

\item{show_leading_edge}{whether to show the leading edge genes in the plot. Default is TRUE}
}
\description{
Plotting drug function on cell type based on the drug perturbed gene set.
}
