% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{rank_celltype}
\alias{rank_celltype}
\title{Prioritizing cell type responsive to drug perturbation.}
\usage{
rank_celltype(
  object,
  n.core = NULL,
  n_dim = 2,
  n_hop = 2,
  perturbed_target = NULL
)
}
\arguments{
\item{object}{scRank object generated from \code{\link{CreateScRank}}.}

\item{n.core}{Number of CPU cores to use. Default is all cores - 1.}

\item{n_dim}{number of low dimensions in manifold space using manifold alignment. Default is 2.}

\item{n_hop}{the number of hop for network propagation. Default is 2.}

\item{perturbed_target}{Target to be perturbed in dpGRN for ranking cell type. Default is the target in \code{object@para$target}}
}
\value{
scRank object containing the drug-responsive cell type rank.
}
\description{
Rank drug-responsive cell type based on perturbation score of drug using manifold alignment and network propagation.
}
