% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{scRank_GSEA}
\alias{scRank_GSEA}
\title{Gene set enrichement analysis for determing the disease relevance of cell type predicted by scRank.}
\usage{
scRank_GSEA(object, celltype = NULL, disease_gene = NULL, selectGenes = NULL)
}
\arguments{
\item{object}{scRank object generated from \code{\link{Constr_net}}}

\item{celltype}{select a cell type to measure its relevance to the disease.}

\item{disease_gene}{a vector containing the positive differential expressed gene names between health and disease state.}

\item{selectGenes}{a character vector containing the selected genes shown in the plot. default is the top 10 genes.}
}
\value{
GSEA plotting with p value indicating the significance of the disease relevance
}
\description{
Use the ranked genes in the cell type and the significantly up-regulated genes in disease condition compared to the normal condition to perform the GSEA analysis. The final p value would reflect the significance of the disease relevance for the cell type.
}
