% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{methods}
\name{show,scRank-method}
\alias{show,scRank-method}
\title{Show scRank object}
\usage{
\S4method{show}{scRank}(object)
}
\arguments{
\item{object}{A \code{scRank} object.}
}
\description{
Show scRank object
}
