% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{subsetScRank}
\alias{subsetScRank}
\title{Subset ScRank Object by Expression}
\usage{
subsetScRank(object, expression, ...)
}
\arguments{
\item{object}{ScRank object}

\item{expression}{Logical expression to subset cells, e.g. condition == 'group1', 'condition' must exist in the column of metadata}

\item{...}{Additional parameters to subset}
}
\description{
Subset ScRank Object by Expression
}
